Remove shared directory. Update each SDK to self contain all logic.
