+++ packages/sdk-react-native/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@orga-ai/sdk-react-native",
-  "version": "0.0.0-test.4",
+  "version": "0.0.0-test.5",
   "main": "lib/index.js",
   "publishConfig": {
     "access": "restricted"
