+++ packages/sdk-react-native/src/index.ts
@@ -1,6 +1,6 @@
-export * from "../../shared";
-export {OrgaAI, ORGAAI_MODELS, ORGAAI_VOICES } from "../../shared"
+export { OrgaAI } from "./core/OrgaAI";
 export {useOrgaAI } from "./hooks/useOrgaAI";
-export { OrgaAIProvider, useOrgaAIContext } from "./hooks/OrgaAIProvider";
+export * from "./hooks/OrgaAIProvider";
 export * from "./types";
+export * from "./errors";
 export { OrgaAICameraView } from "./components/OrgaAICameraView";
