+++ packages/sdk-web/package.json
@@ -1,6 +1,6 @@
 {
-  "name": "orga-ai-sdk",
-  "version": "1.0.0",
+  "name": "@orga-ai/sdk-web",
+  "version": "0.0.0-test.1",
   "main": "lib/index.js",
   "scripts": {
     "build": "tsc --project tsconfig.json",
