--- packages/shared/package.json
@@ -1,28 +1,0 @@
-{
-  "name": "@orga-ai/shared",
-  "version": "0.0.1",
-  "main": "lib/index.js",
-  "types": "lib/index.d.ts",
-  "private": true,
-  "scripts": {
-    "build": "tsc --project tsconfig.json",
-    "clean": "rm -rf dist lib build *.tsbuildinfo",
-    "type-check": "tsc --noEmit --pretty",
-    "test": "jest"
-  },
-  "license": "ISC",
-  "description": "Shared utilities and types for orga-ai-sdk",
-  "dependencies": {
-    "axios": "^1.6.0"
-  },
-  "peerDependencies": {
-    "react": "^18.0.0"
-  },
-  "devDependencies": {
-    "@types/react": "^19.1.8",
-    "jest": "^29.0.0",
-    "react": "^18.0.0",
-    "ts-jest": "^29.0.0",
-    "typescript": "^5.0.0"
-  }
-}
