--- packages/shared/src/index.ts
@@ -1,5 +1,0 @@
-export * from "./core/OrgaAI";
-export * from "./errors";
-export * from "./hooks/OrgaAIProvider";
-export * from "./types";
-export * from "./utils";
